requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-bower==5.2.0
django-filter==2.4.0
django-jsonfield==1.4.1
django-nvd3==0.9.7
django-pandas==0.6.2
djcacheutils==3.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.5
openpyxl==3.0.6
Pillow==8.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
python-memcached==1.59
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2020.5
requests==2.18.3
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.2
whitenoise==5.2.0
xlrd==2.0.1
XlsxWriter==1.3.7
